+++ constants/LightThemeColors.json
@@ -1,0 +1,48 @@
+{
+  "colors": {
+    "primary": "rgb(152, 203, 255)",
+    "onPrimary": "rgb(0, 51, 84)",
+    "primaryContainer": "rgb(0, 74, 119)",
+    "onPrimaryContainer": "rgb(206, 229, 255)",
+    "secondary": "rgb(240, 193, 0)",
+    "onSecondary": "rgb(61, 47, 0)",
+    "secondaryContainer": "rgb(87, 69, 0)",
+    "onSecondaryContainer": "rgb(255, 224, 137)",
+    "tertiary": "rgb(170, 199, 255)",
+    "onTertiary": "rgb(0, 47, 100)",
+    "tertiaryContainer": "rgb(0, 70, 141)",
+    "onTertiaryContainer": "rgb(214, 227, 255)",
+    "error": "rgb(255, 180, 171)",
+    "onError": "rgb(105, 0, 5)",
+    "errorContainer": "rgb(147, 0, 10)",
+    "onErrorContainer": "rgb(255, 180, 171)",
+    "background": "rgb(26, 28, 30)",
+    "onBackground": "rgb(226, 226, 229)",
+    "surface": "rgb(26, 28, 30)",
+    "onSurface": "rgb(226, 226, 229)",
+    "surfaceVariant": "rgb(66, 71, 78)",
+    "onSurfaceVariant": "rgb(194, 199, 207)",
+    "outline": "rgb(140, 145, 153)",
+    "outlineVariant": "rgb(66, 71, 78)",
+    "shadow": "rgb(0, 0, 0)",
+    "scrim": "rgb(0, 0, 0)",
+    "inverseSurface": "rgb(226, 226, 229)",
+    "inverseOnSurface": "rgb(47, 48, 51)",
+    "inversePrimary": "rgb(0, 99, 156)",
+    "elevation": {
+      "level0": "transparent",
+      "level1": "rgb(32, 37, 41)",
+      "level2": "rgb(36, 42, 48)",
+      "level3": "rgb(40, 47, 55)",
+      "level4": "rgb(41, 49, 57)",
+      "level5": "rgb(44, 53, 62)"
+    },
+    "surfaceDisabled": "rgba(226, 226, 229, 0.12)",
+    "onSurfaceDisabled": "rgba(226, 226, 229, 0.38)",
+    "backdrop": "rgba(44, 49, 55, 0.4)",
+    "quaternary": "rgb(247, 189, 67)",
+    "onQuaternary": "rgb(65, 45, 0)",
+    "quaternaryContainer": "rgb(93, 66, 0)",
+    "onQuaternaryContainer": "rgb(255, 222, 165)"
+  }
+}
