+++ stores/useColorTheme.ts
@@ -1,0 +1,14 @@
+import { create } from "zustand";
+
+type ThemeState = {
+  colorTheme: "light" | "dark";
+  toggleTheme: () => void;
+};
+
+export const useColorTheme = create<ThemeState>((set) => ({
+  colorTheme: "light", // Default theme
+  toggleTheme: () =>
+    set((state) => ({
+      colorTheme: state.colorTheme === "light" ? "dark" : "light",
+    })),
+}));
